Number validations also work for float values.
Here's where it becomes important to be able to declare gt and not just ge. As with it you can require, for example, that a value must be greater than 0, even if it is less than 1.
So, 0.5 would be a valid value. But 0.0 or 0 would not.
And the same for lt.